+++ src/app/(blog)/tag/[slug]/page/[n]/page.tsx
@@ -6,7 +6,7 @@ import { t } from '@/lib/i18n'
 import { BlogListing } from '@/components/blog/BlogListing'
 import { parsePathPage } from '@/lib/paginate'
 
-export const dynamic = 'force-dynamic'
+export const revalidate = 3600 // ISR; admin save purges via revalidatePath('/','layout')
 
 export default async function TagPaged({ params }: PageProps<'/tag/[slug]/page/[n]'>) {
   const { slug, n } = await params